+++ internal/exec/runner.go
@@ -37,9 +37,12 @@ import (
 	"text/tabwriter"
 	"time"
 
+	"github.com/golang/protobuf/jsonpb"
+	"github.com/golang/protobuf/proto"
 	"github.com/uber/prototool/internal/breaking"
 	"github.com/uber/prototool/internal/cfginit"
 	"github.com/uber/prototool/internal/create"
+	"github.com/uber/prototool/internal/desc"
 	"github.com/uber/prototool/internal/diff"
 	"github.com/uber/prototool/internal/extract"
 	"github.com/uber/prototool/internal/file"
@@ -52,8 +55,11 @@ import (
 	"github.com/uber/prototool/internal/settings"
 	"github.com/uber/prototool/internal/text"
 	"github.com/uber/prototool/internal/vars"
+	"go.uber.org/multierr"
 	"go.uber.org/zap"
 )
+
+var jsonpbMarshaler = &jsonpb.Marshaler{}
 
 type runner struct {
 	protoSetProvider file.ProtoSetProvider
@@ -247,11 +253,30 @@ func (r *runner) Gen(args []string, dryRun bool) error {
 	return err
 }
 
-func (r *runner) compile(doGen, doFileDescriptorSet, dryRun bool, meta *meta) (protoc.FileDescriptorSets, error) {
+func (r *runner) compile(doGen bool, doFileDescriptorSet bool, dryRun bool, meta *meta) (protoc.FileDescriptorSets, error) {
 	if dryRun {
-		return nil, r.printCommands(doGen, meta.ProtoSet)
+		doFileDescriptorSet = false
 	}
-	compileResult, err := r.newCompiler(doGen, doFileDescriptorSet).Compile(meta.ProtoSet)
+	compiler, err := r.newCompiler(doGen, doFileDescriptorSet, false, false, false)
+	if err != nil {
+		return nil, err
+	}
+	if dryRun {
+		return nil, r.doProtocCommands(compiler, meta)
+	}
+	return r.doCompile(compiler, meta)
+}
+
+func (r *runner) compileFullControl(includeImports bool, includeSourceInfo bool, meta *meta) (protoc.FileDescriptorSets, error) {
+	compiler, err := r.newCompiler(false, false, true, includeImports, includeSourceInfo)
+	if err != nil {
+		return nil, err
+	}
+	return r.doCompile(compiler, meta)
+}
+
+func (r *runner) doCompile(compiler protoc.Compiler, meta *meta) (protoc.FileDescriptorSets, error) {
+	compileResult, err := compiler.Compile(meta.ProtoSet)
 	if err != nil {
 		return nil, err
 	}
@@ -261,12 +286,11 @@ func (r *runner) compile(doGen, doFileDescriptorSet, dryRun bool, meta *meta) (p
 	if len(compileResult.Failures) > 0 {
 		return nil, newExitErrorf(255, "")
 	}
-	r.logger.Debug("protoc command exited without errors")
 	return compileResult.FileDescriptorSets, nil
 }
 
-func (r *runner) printCommands(doGen bool, protoSet *file.ProtoSet) error {
-	commands, err := r.newCompiler(doGen, false).ProtocCommands(protoSet)
+func (r *runner) doProtocCommands(compiler protoc.Compiler, meta *meta) error {
+	commands, err := compiler.ProtocCommands(meta.ProtoSet)
 	if err != nil {
 		return err
 	}
@@ -565,6 +589,61 @@ func (r *runner) GRPC(args, headers []string, address, method, data, callTimeout
 	).Invoke(fileDescriptorSets.Unwrap(), address, method, reader, r.output)
 }
 
+func (r *runner) DescriptorSet(args []string, includeImports bool, includeSourceInfo bool, outputPath string, tmp bool) (retErr error) {
+	if outputPath != "" && tmp {
+		return newExitErrorf(255, "can only set one of output-path, tmp")
+	}
+	meta, err := r.getMeta(args)
+	if err != nil {
+		return err
+	}
+	r.printAffectedFiles(meta)
+	fileDescriptorSets, err := r.compileFullControl(includeImports, includeSourceInfo, meta)
+	if err != nil {
+		return err
+	}
+	fileDescriptorSet, err := desc.MergeFileDescriptorSets(fileDescriptorSets.Unwrap())
+	if err != nil {
+		return err
+	}
+	var data []byte
+	if r.json {
+		buffer := bytes.NewBuffer(nil)
+		err = jsonpbMarshaler.Marshal(buffer, fileDescriptorSet)
+		data = buffer.Bytes()
+	} else {
+		data, err = proto.Marshal(fileDescriptorSet)
+	}
+	if err != nil {
+		return err
+	}
+	if outputPath == "" && !tmp {
+		_, err := r.output.Write(data)
+		return err
+	}
+	var file *os.File
+	if outputPath != "" {
+		file, err = os.Create(outputPath)
+	} else { // if tmp
+		file, err = ioutil.TempFile("", "prototool")
+	}
+	if err != nil {
+		return err
+	}
+	defer func() {
+		retErr = multierr.Append(retErr, file.Close())
+	}()
+	if _, err := file.Write(data); err != nil {
+		return err
+	}
+	if tmp {
+		if err := r.println(file.Name()); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 func (r *runner) InspectPackages(args []string) error {
 	packageSet, _, err := r.getPackageSetAndConfig(args)
 	if err != nil {
@@ -747,7 +826,21 @@ func (r *runner) newDownloader(config settings.Config) (protoc.Downloader, error
 	return protoc.NewDownloader(config, downloaderOptions...)
 }
 
-func (r *runner) newCompiler(doGen bool, doFileDescriptorSet bool) protoc.Compiler {
+func (r *runner) newCompiler(
+	doGen bool,
+	doFileDescriptorSet bool,
+	doFileDescriptorSetFullControl bool,
+	includeImports bool,
+	includeSourceInfo bool,
+) (protoc.Compiler, error) {
+	if doFileDescriptorSet && doFileDescriptorSetFullControl {
+		return nil, fmt.Errorf("cannot set doFileDescriptorSet and doFileDescriptorSetFullControl")
+	}
+	if !doFileDescriptorSetFullControl {
+		if includeImports || includeSourceInfo {
+			return nil, fmt.Errorf("cannot set includeImports or includeSourceInfo without doFileDescriptorSetFullControl")
+		}
+	}
 	compilerOptions := []protoc.CompilerOption{
 		protoc.CompilerWithLogger(r.logger),
 	}
@@ -787,7 +880,13 @@ func (r *runner) newCompiler(doGen bool, doFileDescriptorSet bool) protoc.Compil
 			protoc.CompilerWithFileDescriptorSet(),
 		)
 	}
-	return protoc.NewCompiler(compilerOptions...)
+	if doFileDescriptorSetFullControl {
+		compilerOptions = append(
+			compilerOptions,
+			protoc.CompilerWithFileDescriptorSetFullControl(includeImports, includeSourceInfo),
+		)
+	}
+	return protoc.NewCompiler(compilerOptions...), nil
 }
 
 func (r *runner) newLintRunner() lint.Runner {